# Conflicts:
#	Modules/Social/Http/Livewire/Components/Teams/TeamCalendarList.php
#	Modules/Social/Http/Livewire/Home.php
#	Modules/Social/Http/Livewire/Pages/Teams/Map.php
#	Modules/Social/Resources/views/livewire/components/teams/team-calendar-list.blade.php
#	Modules/Social/Resources/views/livewire/pages/home.blade.php
#	Modules/Social/Resources/views/livewire/pages/teams/calendar.blade.php
#	Modules/Social/Resources/views/livewire/pages/teams/map.blade.php
#	app/Models/Team.php
#	resources/views/livewire/pages/teams/discover.blade.php

+++ Modules/Social/Http/Livewire/DeletePostModal.php
@@ -2,6 +2,7 @@
 
 namespace Modules\Social\Http\Livewire;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Modules\Social\Models\Post;
 use OmniaDigital\OmniaLibrary\Livewire\WithModal;
@@ -9,7 +10,7 @@ use OmniaDigital\OmniaLibrary\Livewire\WithNotification;
 
 class DeletePostModal extends Component
 {
-    use WithModal, WithNotification;
+    use WithModal, WithNotification, AuthorizesRequests;
 
     public Post $post;
 
@@ -29,6 +30,8 @@ class DeletePostModal extends Component
 
     public function deletePost()
     {
+        $this->authorize('delete', $this->post);
+
         if ($this->post) {
             $this->post->delete();
             $this->success('Post deleted successfully');
